Added print formatter for printing cards out in sheets. Slight project restructure
