+++ requirements.txt
@@ -3,3 +3,4 @@ requests == 2.28.1
 pillow == 9.1.0
 multipledispatch == 0.6.0
 blend_modes == 2.1.0
+numpy==1.24.0
